添加插入一条记录功能

+++ Hard/app/src/main/java/com/onecalf/hard/buck/IBaseDao.java
@@ -7,7 +7,7 @@ public interface IBaseDao<T> {
      * @param entity
      * @return
      */
-    int insert(T entity);
+    long insert(T entity);
 
     /**
      * 更新
